requirements.txt
torch
torchvision
opencv-python
matplotlib
Pillow
numpy==1.26  # numpy 2.0 not compatible at the moment
scipy
scikit-learn
faiss-gpu-cu12
tqdm
tifffile
argparse
PyYAML
pandas